+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 ### Changed
+* `SpeedAndDirection` packet now uses 28 speed steps
+* STM32F103 example is now a speed controller
 ### Deprecated
 ### Removed
 ### Fixed